@@ -19,11 +19,11 @@
         }
     ],
     "require": {
-        "php" : "^7.2|^8.0",
-        "symfony/process": "^4.2|^5.0"
+        "php" : "^8.0",
+        "symfony/process": "^5.0"
     },
     "require-dev": {
-        "phpunit/phpunit": "^7.0|^8.0|^9.0"
+        "phpunit/phpunit": "^9.5"
     },
     "autoload": {
         "psr-4": {
