@@ -7,14 +7,7 @@ use Symfony\Component\Process\Process;
 
 class Sqlite extends DbDumper
 {
-    /**
-     * Dump the contents of the database to a given file.
-     *
-     * @param string $dumpFile
-     *
-     * @throws \Spatie\DbDumper\Exceptions\DumpFailed
-     */
-    public function dumpToFile(string $dumpFile)
+    public function dumpToFile(string $dumpFile): void
     {
         $process = $this->getProcess($dumpFile);
 
@@ -23,13 +16,6 @@ class Sqlite extends DbDumper
         $this->checkIfDumpWasSuccessFul($process, $dumpFile);
     }
 
-    /**
-     * Get the command that should be performed to dump the database.
-     *
-     * @param string $dumpFile
-     *
-     * @return string
-     */
     public function getDumpCommand(string $dumpFile): string
     {
         $dumpInSqlite = "echo 'BEGIN IMMEDIATE;\n.dump'";
@@ -47,10 +33,6 @@ class Sqlite extends DbDumper
         return $this->echoToFile($command, $dumpFile);
     }
 
-    /**
-     * @param string $dumpFile
-     * @return Process
-     */
     public function getProcess(string $dumpFile): Process
     {
         $command = $this->getDumpCommand($dumpFile);
