@@ -6,13 +6,7 @@ use Exception;
 
 class CannotSetParameter extends Exception
 {
-    /**
-     * @param string $name
-     * @param string $conflictName
-     *
-     * @return \Spatie\DbDumper\Exceptions\CannotSetParameter
-     */
-    public static function conflictingParameters($name, $conflictName)
+    public static function conflictingParameters(string $name, string $conflictName): static
     {
         return new static("Cannot set `{$name}` because it conflicts with parameter `{$conflictName}`.");
     }
