@@ -6,12 +6,7 @@ use Exception;
 
 class CannotStartDump extends Exception
 {
-    /**
-     * @param string $name
-     *
-     * @return \Spatie\DbDumper\Exceptions\CannotStartDump
-     */
-    public static function emptyParameter($name)
+    public static function emptyParameter(string $name): static
     {
         return new static("Parameter `{$name}` cannot be empty.");
     }
