@@ -7,28 +7,17 @@ use Symfony\Component\Process\Process;
 
 class DumpFailed extends Exception
 {
-    /**
-     * @param \Symfony\Component\Process\Process $process
-     *
-     * @return \Spatie\DbDumper\Exceptions\DumpFailed
-     */
-    public static function processDidNotEndSuccessfully(Process $process)
+    public static function processDidNotEndSuccessfully(Process $process): static
     {
         return new static("The dump process failed with exitcode {$process->getExitCode()} : {$process->getExitCodeText()} : {$process->getErrorOutput()}");
     }
 
-    /**
-     * @return \Spatie\DbDumper\Exceptions\DumpFailed
-     */
-    public static function dumpfileWasNotCreated()
+    public static function dumpfileWasNotCreated(): static
     {
         return new static('The dumpfile could not be created');
     }
 
-    /**
-     * @return \Spatie\DbDumper\Exceptions\DumpFailed
-     */
-    public static function dumpfileWasEmpty()
+    public static function dumpfileWasEmpty(): static
     {
         return new static('The created dumpfile is empty');
     }
